Separate logging and new error type, add docs
