+++ package.json
@@ -38,7 +38,7 @@
     "jest": "^25.1.0",
     "nock": "^12.0.2",
     "supertest": "^4.0.2",
-    "typescript": "^3.7.5"
+    "typescript": "^3.8.3"
   },
   "husky": {
     "hooks": {
