--- src/is_production.js
@@ -1,1 +1,0 @@
-module.exports = () => process.env.NODE_ENV === 'production'
