+++ src/logging/is_logging_enabled.js
@@ -1,0 +1,4 @@
+/**
+ * @returns {boolean} true if NODE_ENV is 'production', false otherwise
+ */
+module.exports = () => process.env.NODE_ENV === 'production'
